v0.5.0: int64 integers, typed buffers, #{} set literals
- Value.as.integer: int32_t → int64_t (full 64-bit range, zero cost)
- Arithmetic/bitwise: all operate on int64 (no more 32-bit wrapping)
- Compiler: integer literals up to INT64_MAX stored natively
- Print: fast int_to_cstr handles 19-digit numbers
- Typed buffers: Int8Array, Int16Array, Int32Array, Uint8Array,
  Uint16Array, Uint32Array, Float32Array, Float64Array
  - Constructor: TypeArray(size) or TypeArray([values])
  - Index access: buf[i] get/set with bounds checking
  - Methods: .len(), .fill(val), .byte_len()
  - GETINDEX returns int64 for integer types, float for float types
- Set literals: #{1, 2, 3} syntax via TOK_SET_LBRACE
  - Methods: .add(), .has(), .delete(), .size(), .clear(), .values()
- 378 tests passing, 0 regressions
- Performance: 10M loop 94ms, 10M sum 165ms (correct results now)

+++ src/value.h
@@ -37,7 +37,7 @@ namespace zen
         union
         {
             bool boolean;
-            int32_t integer;
+            int64_t integer;
             double number;
             Obj *obj;
         } as;
@@ -57,7 +57,7 @@ namespace zen
         v.as.boolean = b;
         return v;
     }
-    inline Value val_int(int32_t i)
+    inline Value val_int(int64_t i)
     {
         Value v;
         v.type = VAL_INT;
